# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amenly",
    "description": "Amenly AI GRC Platform - Production API",
    "version": "1.0.1"
  },
  "paths": {
...
    "/api/v1/organizations/me": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get My Organization",
        "description": "Get current user's organization details including departments and positions.",
        "operationId": "get_my_organization_api_v1_organizations_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "login": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Organizations"
        ],
        "summary": "Update My Organization",
        "description": "Update organization settings (Admin only).",
        "operationId": "update_my_organization_api_v1_organizations_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "login": []
          }
        ]
      }
    },
    "/api/v1/organizations/departments": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "List Departments",
        "description": "List all departments in the current organization.",
        "operationId": "list_departments_api_v1_organizations_departments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "login": []
          }
        ]
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create Department",
        "description": "Create a new department in the organization (Admin only).",
        "operationId": "create_department_api_v1_organizations_departments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "login": []
          }
        ]
      }
    },
    "/api/v1/organizations/positions": {
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create Position",
        "description": "Create a new position within a department (Admin only).",
        "operationId": "create_position_api_v1_organizations_positions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "login": []
          }
        ]
      }
    },
    "/api/v1/organizations/departments/{dept_id}/positions": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "List Positions",
        "description": "List all positions within a specific department.",
        "operationId": "list_positions_api_v1_organizations_departments__dept_id__positions_get",
        "security": [
          {
            "login": []
          }
        ],
        "parameters": [
          {
            "name": "dept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dept Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/departments/{dept_id}": {
      "delete": {
        "tags": [
          "Organizations"
        ],
        "summary": "Delete Department",
        "description": "Delete a department (Admin only).",
        "operationId": "delete_department_api_v1_organizations_departments__dept_id__delete",
        "security": [
          {
            "login": []
          }
        ],
        "parameters": [
          {
            "name": "dept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dept Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/positions/{pos_id}": {
      "delete": {
        "tags": [
          "Organizations"
        ],
        "summary": "Delete Position",
        "description": "Delete a position (Admin only).",
        "operationId": "delete_position_api_v1_organizations_positions__pos_id__delete",
        "security": [
          {
            "login": []
          }
        ],
        "parameters": [
          {
            "name": "pos_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pos Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DepartmentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "organization_id"
        ],
        "title": "DepartmentCreate"
      },
...
      "GenericResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "GenericResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OrganizationUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "OrganizationUpdate"
      },
...
      "PositionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "department_id": {
            "type": "string",
            "format": "uuid",
            "title": "Department Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "department_id"
        ],
        "title": "PositionCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "login": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}